pinctrl: scmi: change `num_pins` type to ensure natural alignment
Change the `num_pins` field's type to `uint32_t` to ensure natural
alignment for C structs that represent this message without using
the pack directive.

Issue: ALB-9426

+++ drivers/pinctrl/nxp/pinctrl-scmi.c
@@ -392,7 +392,7 @@ static int scmi_pinctrl_set_mux_chunk(struct udevice *scmi_dev, u16 no_pins,
 		u16 func;
 	};
 	struct {
-		u8 num_pins;
+		u32 num_pins;
 		struct pin_function pf[];
 	} *request = (void *)buffer;
 	struct {
